Transform thrown exceptions into rejections for all/race.
This is only done for exceptions stemming from bad arguments, i.e. from passing non-iterables, from passing iterables which throw upon getting or calling next, or passing iterables containing objects which cause ToPromise to fail. It is not done for calling the methods with bad `this` values.

See #55 for discussion of this overall theme.