Fix Promise.cast; make Promise.all/race more generic.
`Promise.cast` was temporarily referencing a nonexistent ToPromise as a result of the refactor. I inlined the previous text into its definition.

Then, inside `Promise.all` and `Promise.race`, I now use Invoke to invoke the "cast" method of the called-on constructor, and to invoke the "then" method of the returned-from-cast promise. This latter step closes #53.